+++ backend/routes/user-services.js
@@ -1,62 +1,4 @@
-const express = require("express");
-const router = express.Router();
 const Users = require("../models/users");
-
-router.get("/", async (req, res) => {
-  let id;
-  try {
-    const result = await getUsers(id);
-    res.status(200).send({ users_list: result });
-  } catch (error) {
-    console.log(error);
-    res.status(505).send("An error occurred in the server");
-  }
-});
-
-router.get("/:id", async (req, res) => {
-  const id = req.params["id"];
-  try {
-    const result = await getUsers(id);
-    res.status(200).send({ users_list: result });
-  } catch (error) {
-    console.log(error);
-    res.status(505).send("An error occurred on the system");
-  }
-});
-
-router.post("/", async (req, res) => {
-  const userToAdd = req.body;
-  const result = await addUser(userToAdd);
-  if (result) {
-    res.status(201).send(result);
-  } else {
-    res.status(500).end();
-  }
-});
-
-router.patch("/:id", async (req, res) => {
-  const id = req.params["id"];
-  const username = req.body.username;
-  const first = req.body.first_name;
-  const last = req.body.last_name;
-  try {
-    const result = await updateUserById(id, username, first, last);
-    res.status(200).send({ users_list: result });
-  } catch (error) {
-    console.log(error);
-    res.status(505).send("An error occurred on the system");
-  }
-});
-
-router.delete("/:id", async (req, res) => {
-  const id = req.params["id"];
-  const result = await deleteUserById(id);
-  if (!result) {
-    res.status(404).send("Resource not found");
-  } else {
-    res.status(204).end();
-  }
-});
 
 async function getUsers(id) {
   let result;
@@ -114,10 +56,7 @@ async function deleteUserById(id) {
   }
 }
 
-module.exports = {
-  router,
-  getUsers,
-  addUser,
-  updateUserById,
-  deleteUserById,
-};
+exports.getUsers = getUsers;
+exports.addUser = addUser;
+exports.updateUserById = updateUserById;
+exports.deleteUserById = deleteUserById;
